fix duplicate bot name in status update message (#458)
* fix duplicate bot name in status update message

This commit changes the behavior of the _send_status_update method. If the generic_api is in use, it will avoid adding the bot name to the start of the message because the bot name is already added by the generic api.

* fix duplicate bot name in status update message

This commit moves the fix to the DiscordEmbeds class instead of performing it with an if statement in the GameStats class.