Imagine: being able to from health_manager import mp, hp in your bot!
This PR makes the health_manager effectively into a singleton, which maintains a single global state, rather than keeping it as a class that can be instantiated multiple times.  Unless there is a use case for having multiple health managers running simultaneously in the same process, there's no reason to allow it to instantiate multiple versions of its state.  There are multiple patterns in python to do this, I could have decided on an actual singleton class decorator implementation, or used the "borg" pattern to set the state of all health_managers the same, but this seemed by far the most idiomatic and least verbose.  I've attempted to somewhat limit the changes here, though I've got tabs configured for this project in my editor from editing the ranged bot and other files - so... LMK if you want those back as spaces

The reason I wanted this in the first place was because I'd love to be able to validate that a spell was cast from within my bot by ensuring the mana went down.  There currently seems to be no way to cleanly check the HP/MP from within a bot.  This will allow any bot to simply `from health_manager import hp, mp` to call `hp()` and `mp()` from within their bot class.  Though... y'all might be thinking about my logic for changing this from being a class (no two instantiated in a single process), and realizing the same holds true for a bot, I don't see any reason to prioritize refactoring that, and inheritance can be a useful tool for bot creation.